+++ src/components/admin/VehicleModel.tsx
@@ -156,6 +156,8 @@ const VehicleModel = () => {
   };
 
   const handleDelete = async (id: string) => {
+    const confirmDelete = window.confirm("Are you sure you want to delete?");
+  if (!confirmDelete) return;
     try {
       const response = await fetch(`${API_BASE_URL}/supplier/DeleteVehicleModel/${id}`, { 
         method: "DELETE"
